Move grunt tasks and config into external JS files
