+++ grunt/config/browserify.js
@@ -1,0 +1,7 @@
+module.exports = {
+	commonjs: {
+		files: {
+			'test/commonjs-bundle.js': ['test/commonjs-test.js']
+		}
+	}
+};
