+++ grunt/config/clean.js
@@ -1,0 +1,5 @@
+module.exports = {
+	dist: ['dist'],
+	zipsrc: ['dist/fuelux'],
+	screenshots: ['page-at-timeout-*.jpg']
+};
