+++ grunt/config/compress.js
@@ -1,0 +1,13 @@
+module.exports = {
+	zip: {
+		files: [{
+			cwd: 'dist/',
+			expand: true,
+			src: ['fuelux/**']
+		}],
+		options: {
+			archive: 'dist/fuelux.zip',
+			mode: 'zip'
+		}
+	}
+};
