+++ grunt/config/jsbeautifier.js
@@ -1,0 +1,23 @@
+module.exports = {
+	files: ['dist/js/fuelux.js'],
+	options: {
+		js: {
+			braceStyle: 'collapse',
+			breakChainedMethods: false,
+			e4x: false,
+			evalCode: false,
+			indentLevel: 0,
+			indentSize: 4,
+			indentWithTabs: true,
+			jslintHappy: false,
+			keepArrayIndentation: false,
+			keepFunctionIndentation: false,
+			maxPreserveNewlines: 10,
+			preserveNewlines: true,
+			spaceBeforeConditional: true,
+			spaceInParen: true,
+			unescapeStrings: false,
+			wrapLineLength: 0
+		}
+	}
+};
