+++ grunt/config/uglify.js
@@ -1,0 +1,16 @@
+module.exports = function (grunt) {
+
+	return {
+		options: {
+			report: 'min'
+		},
+		fuelux: {
+			options: {
+				banner: '<%= banner %>'
+			},
+			src: 'dist/js/<%= pkg.name %>.js',
+			dest: 'dist/js/<%= pkg.name %>.min.js'
+		}
+	}
+
+};
